Moved fetching back to newgrid
